Create RecruitUProfilePage2.php
This is the updated version of RecruitUProfilePage.php. It updates the database properly:

Not all input boxes need to be filled in order for the submission to be successful. This allows the user to change absolutely any personal information they wish at any given moment.

Edited the styling, made the font color white and font bold for the name and role.

Stretched out the flashing box so it spans the same row where the name is located at.